requirements.txt
numpy
Pillow
tensorflow-cpu
keras
matplotlib